gui/requirements.txt
gym
pygame
pillow
imageio
imageio-ffmpeg
